+++ lib/auth/auth.ts
@@ -118,7 +118,7 @@ export class Auth {
    *
    * @param opts  Username, password, and an optional set of scopes
    *              When given a set of scopes, they override the default list of
-   *              scopes of `this` intance
+   *              scopes of `this` instance
    *
    * @returns     token object which might contain token(s), scope(s), token type and expiration time
    */