+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Add helpers to split transactions (`models.transaction.createSplit`, `models.transaction.deleteSplit`, `models.transaction.updateSplit`)
+- *Breaking*: renamed `StandingOrderReoccurenceType` to `StandingOrderReoccurrenceType`
+- Add method to fetch account stats (`models.account.getStats`)
 
 ## [0.27.0] - 2020-04-20
 ### Added